+++ lib/src/widgets/router_app/cupertino_vrouter.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/cupertino.dart';
-import 'package:flutter/widgets.dart';
 import 'package:vrouter/src/vlogs.dart';
 
 import 'package:vrouter/src/vrouter_core.dart';
@@ -116,6 +115,7 @@ class CupertinoVRouter extends VRouterApp {
     this.checkerboardOffscreenLayers = false,
     this.showSemanticsDebugger = false,
     this.debugShowCheckedModeBanner = true,
+    this.useInheritedMediaQuery = false,
     this.shortcuts,
     this.actions,
     this.restorationScopeId,
@@ -308,6 +308,9 @@ class CupertinoVRouter extends VRouterApp {
   /// {@macro flutter.widgets.widgetsApp.debugShowCheckedModeBanner}
   final bool debugShowCheckedModeBanner;
 
+  /// {@macro flutter.widgets.widgetsApp.useInheritedMediaQuery}
+  final bool useInheritedMediaQuery;
+
   /// {@macro flutter.widgets.widgetsApp.shortcuts}
   /// {@tool snippet}
   /// This example shows how to add a single shortcut for
@@ -446,6 +449,7 @@ class CupertinoVRouterState extends State<CupertinoVRouter>
         checkerboardOffscreenLayers: widget.checkerboardOffscreenLayers,
         showSemanticsDebugger: widget.showSemanticsDebugger,
         debugShowCheckedModeBanner: widget.debugShowCheckedModeBanner,
+        useInheritedMediaQuery: widget.useInheritedMediaQuery,
         shortcuts: widget.shortcuts,
         actions: widget.actions,
         restorationScopeId: widget.restorationScopeId,